fix: Add more checks to collect_app_status handler

+++ src/events/zookeeper.py
@@ -91,16 +91,17 @@ class ZooKeeperHandler(Object):
             for username, password in internal_user_credentials:
                 self.charm.state.cluster.update({f"{username}-password": password})
 
-        # Kafka keeps a meta.properties in every log.dir with a unique ClusterID
-        # this ID is provided by ZK, and removing it on relation-changed allows
-        # re-joining a ZK cluster undergoing backup restoration.
-        self.charm.workload.exec(
-            [
-                "bash",
-                "-c",
-                f"""find {self.charm.workload.paths.data_path} -type f -name meta.properties -delete || true""",
-            ]
-        )
+        if not self.charm.state.zookeeper.endpoints:
+            # Kafka keeps a meta.properties in every log.dir with a unique ClusterID
+            # this ID is provided by ZK, and removing it on relation-changed allows
+            # re-joining a ZK cluster undergoing backup restoration.
+            self.charm.workload.exec(
+                [
+                    "bash",
+                    "-c",
+                    f"""find {self.charm.workload.paths.data_path} -type f -name meta.properties -delete || true""",
+                ]
+            )
 
         # attempt re-start of Kafka for all units on zookeeper-changed
         # avoids relying on deferred events elsewhere that may not exist after cluster init
